+++ src/common/index.ts
@@ -1,8 +1,0 @@
-import type { ResultType } from "./result";
-
-export { compareArrays } from "./compareArrays";
-import type { FileError } from "./fileError";
-
-export { niceDisplay } from "./niceDisplay";
-
-export { type ResultType, type FileError };