+++ src/features/compareJsonObjects.ts
@@ -1,0 +1,9 @@
+import { Json, ResultType } from "@/lib/types";
+import { validateJson } from "@/lib/jsons";
+
+export function compareJsonObjects(
+  main: Json,
+  rest: Json
+): ResultType<Array<string>> {
+  return validateJson(main, rest);
+}