+++ src/index.ts
@@ -1,5 +1,1 @@
-import { niceDisplay } from "./common";
-import { validateDirs } from "./validateDirs";
-import { validateFiles } from "./validateJson";
-
-export { niceDisplay, validateDirs, validateFiles };
+export * from "@/features";