+++ src/lib/compareArrays/index.ts
@@ -1,0 +1,1 @@
+export { compareArrays } from "./compareArrays";