+++ src/lib/displayErrors/index.ts
@@ -1,0 +1,1 @@
+export { displayErrors } from "./displayErrors";