+++ src/lib/types/index.ts
@@ -1,0 +1,4 @@
+export { type FileError } from "./fileError";
+export { type ResultType } from "./result";
+export { type Json } from "./json";
+export { type FileToValidate } from "./fileToValidate";